move lib/admin under rpc/admin, and some case changes
