+++ mempool/partitioner/partitioner_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/0xsequence/bundler/contracts/gen/solabis/abiendorser"
 	"github.com/0xsequence/bundler/endorser"
-	"github.com/0xsequence/bundler/lib/mempool/partitioner"
 	"github.com/0xsequence/bundler/lib/types"
+	"github.com/0xsequence/bundler/mempool/partitioner"
 	"github.com/0xsequence/ethkit/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
 )
