--- rpc/ops.go
@@ -1,1 +1,0 @@
-package rpc
