+++ sender/sender.go
@@ -10,9 +10,9 @@ import (
 	"github.com/0xsequence/bundler/endorser"
 	"github.com/0xsequence/bundler/lib/collector"
 	"github.com/0xsequence/bundler/lib/interfaces"
-	"github.com/0xsequence/bundler/lib/mempool"
 	"github.com/0xsequence/bundler/lib/registry"
 	"github.com/0xsequence/bundler/lib/utils"
+	"github.com/0xsequence/bundler/mempool"
 	"github.com/0xsequence/bundler/sender/chiller"
 	"github.com/0xsequence/bundler/sender/worker"
 	"github.com/go-chi/httplog/v2"
