+++ sender/sender_test.go
@@ -12,11 +12,11 @@ import (
 	"github.com/0xsequence/bundler/contracts/gen/solabis/abivalidator"
 	"github.com/0xsequence/bundler/endorser"
 	"github.com/0xsequence/bundler/lib/collector"
-	"github.com/0xsequence/bundler/lib/mempool"
 	"github.com/0xsequence/bundler/lib/mocks"
 	"github.com/0xsequence/bundler/lib/pricefeed"
 	"github.com/0xsequence/bundler/lib/registry"
 	"github.com/0xsequence/bundler/lib/types"
+	"github.com/0xsequence/bundler/mempool"
 	"github.com/0xsequence/bundler/proto"
 	"github.com/0xsequence/bundler/sender"
 	"github.com/0xsequence/ethkit/ethtxn"
